requirements.txt

wxPython
accessible_output2
keyring
IMAPClient
yagmail
windows-toasts
pystray
Pillow
